requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
docopt==0.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pipreqs==0.4.13
python-dotenv==1.0.1
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
tornado==6.4.2
urllib3==2.3.0
yarg==0.1.10
bleach==6.1.0
cachetools==5.3.2
cryptography==42.0.0